+++ packages/create-hydrogen-app/CHANGELOG.md
@@ -13,6 +13,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 - feat: `/src/entry-client.jsx` file has been removed. The new entry point in for the client in `index.html` is `/@shopify/hydrogen/entry-client`. Custom entry points are still supported
 - fix: Footer date update
 - fix: product link errors in Cart.client.jsx of the example template
+- feat: Helmet component has been renamed to Head
 
 ## 0.10.1 - 2022-01-26